Verificação de cargo de usuário admin aprimorado. Correção de bugs ao efetuar atualização de foto de perfil do cliente

+++ painel/ajax/forms.php
@@ -17,20 +17,30 @@
         }
     }
     
-    if(isset($_GET['signup']) && $_GET['signup'] == 'signup-client'){
+    if(isset($_POST['signup']) && $_POST['signup'] == 'signup-client'){
         sleep(1.5);
-        $nome = $_GET['nome'];
-        $email = $_GET['email'];
-        $senha = $_GET['password'];
-        $senha2 =  $_GET['confirmPassword'];
+        $nome = $_POST['nome'];
+        $email = $_POST['email'];
+        $senha = $_POST['password'];
+        $senha2 =  $_POST['confirmPassword'];
         $img = "";
+
+        if(isset($_FILES['img'])){
+            if(Painel::imagemValida($_FILES['img'])){
+               $img = $_FILES['img'];
+            }else{
+                $img = "";
+                $data['sucesso'] = false;
+                $data['msg'] = "Arquivo ou tamanho da imagem inválido.";
+            }
+        }
 
         if($nome == ''){
             $data['sucesso'] = false;
             $data['msg'] = "O campo Nome não pode estar vázio.";
         }
 
-        if(isset($_GET['email']) && $email != ''){
+        if(isset($_POST['email']) && $email != ''){
             $sql = MySql::conectar()->prepare("SELECT `id` FROM `tb_site.clientes` WHERE email = ?");
             $sql->execute(array($email));
                 if($sql->rowCount() != 0){
@@ -50,21 +60,11 @@
             $data['msg'] = "As senhas digitadas não coincidem!";
         }
 
-        if(isset($_FILES['img'])){
-            if(Painel::imagemValida($_FILES['img'])){
-                $img = $_FILES['img'];
-            }else{
-                $img = "";
-                $data['sucesso'] = false;
-                $data['msg'] = "Arquivo ou tamanho da imagem inválido.";
-            }
-        }
-
-        if($data['sucesso'] == true){
+        if($data['sucesso']){
             if(is_array($img))
-                $img = Painel::uploadFile($img, 'uploads/clientes');
+                $img = Painel::uploadFile('uploads/clientes', $img);
             \view\Login::signUp($nome, $email, $senha, $img);
-            $data['msg'] = "Cadastrado realizado com sucesso!";
+            $data['msg'] = "Cadastro de cliente realizado com sucesso!";
         }
 
     }else if(isset($_POST['tipo_acao']) && $_POST['tipo_acao'] == 'editar_cliente'){
@@ -92,7 +92,7 @@
 
         if($data['sucesso']){
             if(is_array($img)){
-                $img = Painel::uploadFile($img, 'uploads/clientes');
+                $img = Painel::updateFile($img, 'uploads/clientes');
             }
             $sql = MySql::conectar()->prepare("UPDATE `tb_site.clientes` SET nome = ?, email = ?, senha = ?, img = ? WHERE id = $id");
             $sql->execute(array($nome, $email, $senha, $img));
@@ -101,14 +101,14 @@
         }
 
     }else if(isset($_POST['tipo_acao']) && $_POST['tipo_acao'] == 'excluir_cliente'){
-        $id = $_POST['id'];
-        $sql = MySql::conectar()->prepare("SELECT img FROM `tb_site.clientes` WHERE id = $id");
-        $sql->execute();
-        $img = $sql->fetch()['img'];
-        @unlink('../uploads/clientes/'.$img);
-
-        MySql::conectar()->exec("DELETE FROM `tb_site.clientes` WHERE id = $id");    
-
+        if($_SESSION >= 2){
+            $id = $_POST['id'];
+            $sql = MySql::conectar()->prepare("SELECT img FROM `tb_site.clientes` WHERE id = $id");
+            $sql->execute();
+            $img = $sql->fetch()['img'];
+            @unlink('../uploads/clientes/'.$img);
+            MySql::conectar()->exec("DELETE FROM `tb_site.clientes` WHERE id = $id");
+        }
     }else if(isset($_POST['tipo_acao']) && $_POST['tipo_acao'] == 'ordenar_concessionaria'){
         $ids = $_POST['item'];
         orderId($ids, 'tb_site.concessionarias');
